SFR-2261: Refactoring db migration process
