+++ processes/__init__.py
@@ -13,7 +13,7 @@ from .api import APIProcess
 from .muse import MUSEProcess
 from .ingestReport import IngestReportProcess
 from .met import METProcess
-from .migration import MigrationProcess
+from .db_migration import MigrationProcess
 from .dbMaintenance import DatabaseMaintenanceProcess
 from .chicagoISAC import ChicagoISACProcess
 from .UofSC import UofSCProcess
